cpyext: compile the variadic entry points instead of inlining them
The bodies of `PyArg_Parse*`, `Py_BuildValue`, `PyObject_CallFunction` /
`CallMethod` and their `ObjArgs` pair, `PyUnicode_FromFormat*`,
`PyErr_Format`, `PyErr_FormatUnraisable` and `PyOS_[v]snprintf` move out of
`modsupport.h`, `pyre_format.h`, `abstract.h` and `pyerrors.h` into
`pyre-interpreter/src/cpyext/src/`, the six translation units
`pypy/module/cpyext/src/getargs.c` and its siblings hold.  The headers keep
the declarations; the sources are compiled by `pyre-interpreter/build.rs`
under the predicate `lib.rs` gates `mod cpyext` with.

`cpyext/csources.rs` declares the seventeen entry points and takes their
addresses, which is what pulls each archive member into the binary;
`pyrex/build.rs` reads that same block for the names it exports.
`PyObject_CallFunction` and `PyObject_CallMethod` reach the unit walk
through `Py_VaBuildValue` rather than through the `_PyPyre_BuildValue` that
is now file-local, which is also the spelling `abstract.c` uses upstream.

`cpyext-abi.py check` reads the C definitions the way it reads the header
inlines, so the same twenty-four non-Rust entry points are still measured
against the recorded CPython declarations.

modsupport.h 917 -> 33 lines, pyre_format.h 389 -> 29, abstract.h 172 -> 61.
